nh - Created a controller directory which will hold controller files. Fixed errors that resulted from this change.
