--- src/main/resources/templates/page1.html
@@ -1,19 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org" lang="en"> 
-  <head>
-    <title>Title of your page goes here</title>
-    <div th:replace="bootstrap/bootstrap_head.html"></div>
-  </head>
-  <body>
-    <div class="container">
-      <div th:replace="bootstrap/bootstrap_nav_header.html"></div>
-
-      <h1>Page 1</h1>
-
-      <p>This page is a placeholder.</p>
-
-      <div th:replace="bootstrap/bootstrap_footer.html"></div>
-    </div>
-    <div th:replace="bootstrap/bootstrap_scripts.html"></div>
-  </body>
-</html>
